+++ client_src/js/monitoring/monitoring-view.state.js
@@ -1,6 +1,6 @@
 // Store the current active view type like camera, panel, location map or list
 let currentViewType = "none";
-const viewTypes = ["panel", "list", "camera", "map", "current-ops"];
+const viewTypes = ["panel", "list", "camera", "group", "current-ops", "combined"];
 
 let printerInfo = null;
 let printerControlList = null;