+++ server_src/config/users.js
@@ -2,14 +2,19 @@ const { SettingsClean } = require("../lib/dataFunctions/settingsClean.js");
 
 module.exports = {
   async ensureCurrentUserAndGroup(req, res, next) {
-    const serverSettings = await SettingsClean.returnSystemSettings();
+    const serverSettings = SettingsClean.returnSystemSettings();
+    const clientSettings = SettingsClean.returnClientSettings();
 
     // If login is not required, set default user and admin otherwise pass current user/group.
     if (!serverSettings?.server?.loginRequired) {
       req.user = {
-        name: "No User",
-        group: "Administrator"
+        userID: "BLANK",
+        name: "",
+        group: "Administrator",
+        clientSettings: clientSettings
       };
+    } else {
+      req.user.clientSettings = clientSettings;
     }
 
     next();