+++ server_src/routes/index.js
@@ -20,12 +20,12 @@ const version = process.env[AppConstants.VERSION_KEY];
 
 // Welcome Page
 router.get("/", async (req, res) => {
-  const serverSettings = await ServerSettings.find({});
+  const serverSettings = SettingsClean.returnSystemSettings();
 
-  if (serverSettings[0].server.loginRequired === false) {
+  if (serverSettings.server.loginRequired === false) {
     res.redirect("/dashboard");
   } else {
-    const { registration } = serverSettings[0].server;
+    const { registration } = serverSettings.server;
 
     if (req.isAuthenticated()) {
       res.redirect("/dashboard");
@@ -34,7 +34,7 @@ router.get("/", async (req, res) => {
         page: "Welcome",
         octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
         registration,
-        serverSettings: serverSettings[0]
+        serverSettings: serverSettings
       });
     }
   }
@@ -43,9 +43,9 @@ router.get("/", async (req, res) => {
 // Dashboard Page
 router.get("/dashboard", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
-  const clientSettings = await SettingsClean.returnClientSettings();
-  const dashStatistics = await PrinterClean.returnDashboardStatistics();
-  let dashboardSettings = clientSettings?.dashboard || getDefaultDashboardSettings();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const dashStatistics = PrinterClean.returnDashboardStatistics();
+  let dashboardSettings = req.user.clientSettings?.dashboard || getDefaultDashboardSettings();
 
   res.render("dashboard", {
     name: req.user.name,
@@ -56,12 +56,14 @@ router.get("/dashboard", ensureAuthenticated, ensureCurrentUserAndGroup, async (
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     helpers: prettyHelpers,
     dashboardSettings: dashboardSettings,
-    dashboardStatistics: dashStatistics
+    dashboardStatistics: dashStatistics,
+    serverSettings,
+    clientSettings: req.user.clientSettings
   });
 });
 router.get("/printers", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
-  const serverSettings = await SettingsClean.returnSystemSettings();
+  const serverSettings = SettingsClean.returnSystemSettings();
   res.render("printerManagement", {
     name: req.user.name,
     userGroup: req.user.group,
@@ -70,15 +72,17 @@ router.get("/printers", ensureAuthenticated, ensureCurrentUserAndGroup, async (r
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     printerCount: printers.length,
     helpers: prettyHelpers,
-    air_gapped: softwareUpdateChecker.getUpdateNotificationIfAny().air_gapped
+    air_gapped: softwareUpdateChecker.getUpdateNotificationIfAny().air_gapped,
+    serverSettings,
+    clientSettings: req.user.clientSettings
   });
 });
 // File Manager Page
 router.get("/filemanager", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
-  const printers = await PrinterClean.listPrintersInformation();
-  const serverSettings = await SettingsClean.returnSystemSettings();
-  const currentOperations = await PrinterClean.returnCurrentOperations();
-  const fileStatistics = await FileClean.returnStatistics();
+  const printers = PrinterClean.listPrintersInformation();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const currentOperations = PrinterClean.returnCurrentOperations();
+  const fileStatistics = FileClean.returnStatistics();
   res.render("filemanager", {
     name: req.user.name,
     userGroup: req.user.group,
@@ -89,7 +93,9 @@ router.get("/filemanager", ensureAuthenticated, ensureCurrentUserAndGroup, async
     printers,
     helpers: prettyHelpers,
     currentOperationsCount: currentOperations.count,
-    fileStatistics
+    fileStatistics,
+    serverSettings,
+    clientSettings: req.user.clientSettings
   });
 });
 // History Page
@@ -98,6 +104,7 @@ router.get("/history", ensureAuthenticated, ensureCurrentUserAndGroup, async (re
   const historyCache = getHistoryCache();
   const history = historyCache.historyClean;
   const statistics = historyCache.statisticsClean;
+  const serverSettings = SettingsClean.returnSystemSettings();
 
   res.render("history", {
     name: req.user.name,
@@ -108,20 +115,22 @@ router.get("/history", ensureAuthenticated, ensureCurrentUserAndGroup, async (re
     printStatistics: statistics,
     helpers: prettyHelpers,
     page: "History",
-    octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY]
+    serverSettings,
+    octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
+    clientSettings: req.user.clientSettings
   });
 });
 
 // Panel view  Page
 router.get("/mon/panel", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
-  const sortedIndex = await Runner.sortedIndex();
-  const clientSettings = await SettingsClean.returnClientSettings();
-  const dashStatistics = await PrinterClean.returnDashboardStatistics();
+  const sortedIndex = Runner.sortedIndex();
+  const dashStatistics = PrinterClean.returnDashboardStatistics();
   const currentSort = getSorting();
   const currentFilter = getFilter();
+  const serverSettings = SettingsClean.returnSystemSettings();
 
-  let printGroups = await Runner.returnGroupList();
+  let printGroups = Runner.returnGroupList();
   if (typeof printGroups === "undefined") {
     printGroups = [];
   }
@@ -136,8 +145,9 @@ router.get("/mon/panel", ensureAuthenticated, ensureCurrentUserAndGroup, async (
     page: "Panel View",
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     helpers: prettyHelpers,
-    clientSettings,
+    clientSettings: req.user.clientSettings,
     printGroups,
+    serverSettings,
     currentChanges: { currentSort, currentFilter },
     dashboardStatistics: dashStatistics
   });
@@ -146,9 +156,8 @@ router.get("/mon/panel", ensureAuthenticated, ensureCurrentUserAndGroup, async (
 router.get("/mon/camera", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
   const sortedIndex = await Runner.sortedIndex();
-  const clientSettings = await SettingsClean.returnClientSettings();
-  const serverSettings = await SettingsClean.returnSystemSettings();
-  const dashStatistics = await PrinterClean.returnDashboardStatistics();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const dashStatistics = PrinterClean.returnDashboardStatistics();
   const currentSort = getSorting();
   const currentFilter = getFilter();
 
@@ -167,17 +176,17 @@ router.get("/mon/camera", ensureAuthenticated, ensureCurrentUserAndGroup, async 
     page: "Camera View",
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     helpers: prettyHelpers,
-    clientSettings,
+    clientSettings: req.user.clientSettings,
     printGroups,
+    serverSettings,
     currentChanges: { currentSort, currentFilter },
     dashboardStatistics: dashStatistics
   });
 });
-router.get("/mon/printerMap", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
+router.get("/mon/group", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
   const sortedIndex = await Runner.sortedIndex();
-  const clientSettings = await SettingsClean.returnClientSettings();
-  const serverSettings = await SettingsClean.returnSystemSettings();
+  const serverSettings = SettingsClean.returnSystemSettings();
 
   const currentSort = getSorting();
   const currentFilter = getFilter();
@@ -194,11 +203,12 @@ router.get("/mon/printerMap", ensureAuthenticated, ensureCurrentUserAndGroup, as
     printers,
     printerCount: printers.length,
     sortedIndex,
-    page: "Printer Map",
+    page: "Group",
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     helpers: prettyHelpers,
-    clientSettings,
+    clientSettings: req.user.clientSettings,
     printGroups,
+    serverSettings,
     currentChanges: { currentSort, currentFilter }
   });
 });
@@ -206,12 +216,13 @@ router.get("/mon/printerMap", ensureAuthenticated, ensureCurrentUserAndGroup, as
 router.get("/mon/list", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
   const sortedIndex = await Runner.sortedIndex();
-  const clientSettings = await SettingsClean.returnClientSettings();
-  const dashStatistics = await PrinterClean.returnDashboardStatistics();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const clientSettings = SettingsClean.returnClientSettings();
+  const dashStatistics = PrinterClean.returnDashboardStatistics();
   const currentSort = getSorting();
   const currentFilter = getFilter();
 
-  let printGroups = await Runner.returnGroupList();
+  let printGroups = Runner.returnGroupList();
   if (typeof printGroups === "undefined") {
     printGroups = [];
   }
@@ -228,14 +239,47 @@ router.get("/mon/list", ensureAuthenticated, ensureCurrentUserAndGroup, async (r
     helpers: prettyHelpers,
     clientSettings,
     printGroups,
+    serverSettings,
     currentChanges: { currentSort, currentFilter },
     dashboardStatistics: dashStatistics
   });
 });
+
+router.get("/mon/combined", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
+  const printers = await Runner.returnFarmPrinters();
+  const sortedIndex = await Runner.sortedIndex();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const dashStatistics = PrinterClean.returnDashboardStatistics();
+  const currentSort = getSorting();
+  const currentFilter = getFilter();
+
+  let printGroups = Runner.returnGroupList();
+  if (typeof printGroups === "undefined") {
+    printGroups = [];
+  }
+
+  res.render("combinedView", {
+    name: req.user.name,
+    userGroup: req.user.group,
+    version,
+    printers,
+    printerCount: printers.length,
+    sortedIndex,
+    page: "Combined View",
+    octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
+    helpers: prettyHelpers,
+    clientSettings: req.user.clientSettings,
+    printGroups,
+    serverSettings,
+    currentChanges: { currentSort, currentFilter },
+    dashboardStatistics: dashStatistics
+  });
+});
+
 router.get("/mon/currentOp", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
   const printers = await Runner.returnFarmPrinters();
   const sortedIndex = await Runner.sortedIndex();
-  const clientSettings = await SettingsClean.returnClientSettings();
+  const serverSettings = SettingsClean.returnSystemSettings();
 
   res.render("currentOperationsView", {
     name: req.user.name,
@@ -247,7 +291,8 @@ router.get("/mon/currentOp", ensureAuthenticated, ensureCurrentUserAndGroup, asy
     page: "Current Operations",
     octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
     helpers: prettyHelpers,
-    clientSettings
+    serverSettings,
+    clientSettings: req.user.clientSettings
   });
 });
 router.get("/filament", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
@@ -255,10 +300,10 @@ router.get("/filament", ensureAuthenticated, ensureCurrentUserAndGroup, async (r
   const historyStats = historyCache.generateStatistics();
 
   const printers = Runner.returnFarmPrinters();
-  const serverSettings = await SettingsClean.returnSystemSettings();
-  const statistics = await FilamentClean.getStatistics();
-  const spools = await FilamentClean.getSpools();
-  const profiles = await FilamentClean.getProfiles();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const statistics = FilamentClean.getStatistics();
+  const spools = FilamentClean.getSpools();
+  const profiles = FilamentClean.getProfiles();
 
   res.render("filament", {
     name: req.user.name,
@@ -272,7 +317,8 @@ router.get("/filament", ensureAuthenticated, ensureCurrentUserAndGroup, async (r
     spools,
     profiles,
     statistics,
-    historyStats
+    historyStats,
+    clientSettings: req.user.clientSettings
   });
 });